# Build: Send benchmarking base list to bot